backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
gunicorn==23.0.0
sse-starlette==2.1.0
python-multipart==0.0.12

# Async support
httpx==0.27.2
aiohttp==3.10.10

# Pydantic for schemas
pydantic>=2.11.0
pydantic-settings>=2.5.2

# Azure SDKs
azure-identity==1.19.0
openai>=1.99.0

# Microsoft Agent Framework (core only)
agent-framework-core==1.0.0b260128

# Redis for event streaming
redis==5.2.0

# Database
asyncpg==0.30.0

# Data Source Layer — Azure AI Search, Cosmos DB, Fabric SQL
azure-search-documents>=11.4.0
azure-cosmos>=4.7.0
pyodbc>=5.1.0

# Observability
opentelemetry-api>=1.39.0
opentelemetry-sdk>=1.39.0
opentelemetry-instrumentation-fastapi>=0.50b0
opentelemetry-exporter-otlp>=1.39.0
opentelemetry-semantic-conventions-ai==0.4.13
azure-monitor-opentelemetry>=1.6.3

# Utils
python-dotenv==1.0.1
structlog==24.4.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
